refactor: suggestion 관련 기능 suggestion package 내로 이동
