--- src/main/java/com/beyond/easycheck/infrastructure/persistence/entity/AgreementType.java
@@ -1,7 +1,0 @@
-package com.beyond.easycheck.infrastructure.persistence.entity;
-
-public enum AgreementType {
-
-    Agree,
-    DisAgree
-}
