+++ src/main/java/com/beyond/easycheck/suggestions/infrastructure/persistence/entity/AgreementType.java
@@ -1,0 +1,7 @@
+package com.beyond.easycheck.suggestions.infrastructure.persistence.entity;
+
+public enum AgreementType {
+
+    Agree,
+    DisAgree
+}
